sem2/project 2sem/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.13.1
python-dotenv==1.0.0
opencv-python==4.9.0.80
numpy==1.26.4
redis==5.0.1
celery==5.3.6
email-validator==2.1.0.post1